+++ src/controls/index.ts
@@ -1,4 +1,0 @@
-export * from './control.class';
-export * from './actions';
-export * from './controls.render';
-import './default_controls';