+++ src/shapes/polygon.class.ts
@@ -1,4 +1,3 @@
-import { fabric } from '../../HEADER';
 import { TClassProperties } from '../typedefs';
 import { classRegistry } from '../util/class_registry';
 import { Polyline, polylineDefaultValues } from './polyline.class';
@@ -18,6 +17,3 @@ Object.assign(Polygon.prototype, polygonDefaultValues);
 
 classRegistry.setClass(Polygon);
 classRegistry.setSVGClass(Polygon);
-
-/** @todo TODO_JS_MIGRATION remove next line after refactoring build */
-fabric.Polygon = Polygon;